Added wrapper methods to ProtocolHandler to push events into the respective protocol handler's event.
Removed Launcher Outdated short circuit.

Should be ready for pushing uphill.

----------------------------

Usage Changes:

In the LauncherConfig.ini:
The FTPUsername, FTPPassword, and FTPUrl fields have been changed to PatchUsername, PatchPassword, and PatchUrl respectively.

Method Changes:

In FTPHandler.cs:
Methods ReadFTPFile and both DownloadFTPFile(s) have been renamed to ReadPatchFile and DownloadPatchFile. This was done so that the shim names would be the same in the shim as well as both protocols.