Added stub classes for BitTorrent and HTTP.
This commit marks the start of some sweeping internal changes in the
launcher. Most of the heavy lifting will be moved from the GameHandler
and LauncherHandler classes into the protocol classes instead.
GameHandler and LauncherHandler will, in the future, focus on
retrieving information about the new patch from a central server and
moving downloaded files into the game directory.
The roadmap for this refactoring is as follows:

* Finish design of new system
* Implement FTP protocol in a transparent fashion - the new system
should not require serverside changes for the protocol.
* Implement HTTP/HTTPS protocol under the new system.
* Implement BitTorrent protocol under the new system.